Mount can deadlock, use better signals
So for some reason, if we call IoAttachDevice between the FDO
and diskDeviceObject, we can sometimes deadlock, and always
deadlock if Avast is installed. Unknown why, but
it seems we can continue without attaching.

Change the signal to userland that mount is completed, if
the IOCTL is stalled waiting, the process is halted, including
WorkItem functions.

Turns out Windows does have some userland/kernel signal, so
we now use a named event, and userland process will wait
instead for the kernel to signal it is ready.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>