Correct SecurityDescriptors
When a new file/dir is created, first fish out the
OwnerID/GroupID from the running process, and map
to uid,gid for vattr.

Then after the creation, fish them out again
and create a new SecurityDescriptor, and attach to vnode
and save to _NTACL. This way we keep the original SIDs for
Windows, and have a non-nobody mapping for Unix.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>