Change OpenZVOL to SCSIAdapter
ChatGPT claims it is more proper to be a SCSIAdapter
than attaching as a Volume. This makes those changes,
mostly to .inf and zfsinstaller.exe.

We also remove the attaching to storport device stack, as
it can stop all IRPs from flowing, and ZVOL support stop
working. Now we open the PDO device to hold it busy instead.

Change zfsinstaller to always clean up ghost installs.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>