ReparsePoint math did not account for StreamName
When we come across a ReparsePoint we are to return
the length of the "not-yet-parsed" section of the FileName,
which may have had ":Zone.Identifier" at the end. Now
we will add on the stream name length as well.
Additionally, ":$DATA" can be omitted, if it is we
have to remove that length, as internally we always
add ":$DATA".

Fix BSOD if non-existent streamname was attempted
to be opened, without CREATE flag. VNRELE on NULL.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>