When issuing IO handle broken, missing devices and failures
Previously we would issue the IO and assume everything
is going according to plan, this meant in the case of
failure to issue IO, we would hang forever. For example
if the device is removed.

Now we check the return code of IoCallDriver(), and handle
both direct failures, or sync calls (instead of expected
async call).

This gives us a better chance of detecting ejected, removed
or failed devices.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>